split crawling, edit run_test
